refactor: Organize package by domain
